requirements.txt
clipboard==0.0.4
jira==3.4.1
pandas==2.1.4
python-dotenv=0.21.0
streamlit==1.33.0